+++ osgCigi/World.h
@@ -17,8 +17,8 @@ namespace osgCigi
 		void updateEntityOrientation(int id, osg::Vec3f orientation);
 		void updateEntityAngularVelocity(int id, osg::Vec3f velocity);
 		void updateEntityAcceleration(int id, osg::Vec3f acceleration);
-		void setTimestamp(float t);
-		float getTimestamp() const;
+		void setTimestamp(long t);
+		long getTimestamp() const;
 		void firstOrderPredictUpdate(
 			const std::function<float(float p, float v)>& predictionFunction);
 		//void instantCorrect(World* reference, double threshold);
@@ -26,7 +26,7 @@ namespace osgCigi
 		virtual ~World();
 	private:
 		std::list<Entity> entities;
-		float timestamp;
+		long timestamp;
 		mutable std::mutex m;
 	};
 }
